+++ ecosystem.config.cjs
@@ -7,16 +7,17 @@ module.exports = {
     {
       name: `edge-daemon`,
       script: 'pnpm prod:cli edge daemon serve',
-      cron_restart: '0 0 * * *',
+      // cron_restart: '0 0 * * *',
     },
     {
       name: `edge-ftp`,
       script: 'pnpm prod:cli edge ftp serve',
-      cron_restart: '0 0 * * *',
+      // cron_restart: '0 0 * * *',
     },
     {
-      name: `edge-syslog`,
-      script: 'pnpm prod:cli edge syslog serve',
+      name: `edge-volume`,
+      script: 'pnpm prod:cli edge volume mount',
+      // cron_restart: '0 0 * * *',
     },
     {
       name: `mothership`,