+++ packages/pockethost/src/common/ioc.ts
@@ -1,5 +1,8 @@
 type ServiceMap = {
-  [serviceName: string]: unknown
+  [serviceName: string]: {
+    service: any
+    stack?: string
+  }
 }
 
 const services: ServiceMap = {}
@@ -7,12 +10,15 @@ const services: ServiceMap = {}
 export const ioc = <T = unknown>(serviceName: string, register?: T): T => {
   if (register) {
     if (serviceName in services) {
-      throw new Error(`Service with key '${serviceName}' already registered.`)
+      throw new Error(
+        `Service with key '${serviceName}' already registered. First registered: ${services[serviceName]!.stack}`,
+      )
+    } else {
+      services[serviceName] = { service: register, stack: new Error().stack }
     }
-    services[serviceName] = register
   }
   if (!(serviceName in services)) {
     throw new Error(`Service with key '${serviceName}' not found.`)
   }
-  return services[serviceName] as T
+  return services[serviceName]!.service as T
 }