+++ packages/pockethost/src/core/exit.ts
@@ -1,7 +1,7 @@
 import exitHook, { asyncExitHook as _, gracefulExit as __ } from 'exit-hook'
 import { logger } from '../common'
 
-export const asyncExitHook = (cb: () => Promise<void>) => _(cb, { wait: 5000 })
+export const asyncExitHook = (cb: () => Promise<any>) => _(cb, { wait: 5000 })
 
 export const gracefulExit = async (signal?: number) => {
   __(signal)