+++ packages/pockethost/src/mothership-app/src/lib/handlers/instance/bootstrap/HandleInstancesResetIdle.ts
@@ -1,0 +1,4 @@
+export const HandleInstancesResetIdle = (e: core.BootstrapEvent) => {
+  const dao = $app.dao()
+  dao.db().newQuery(`update instances set status='idle'`).execute()
+}