+++ packages/pockethost/src/mothership-app/src/lib/handlers/stats/hooks.ts
@@ -1,0 +1,3 @@
+routerAdd('GET', '/api/stats', (c) => {
+  return require(`${__hooks}/mothership`).HandleStatsRequest(c)
+})