+++ packages/pockethost/src/mothership-app/src/lib/util/versions.ts
@@ -1,0 +1,1 @@
+export const versions = require(`${__hooks}/versions.cjs`)