+++ packages/pockethost/CHANGELOG.md
@@ -1,5 +1,29 @@
 # pockethost
 
-## 2.0.0-rc.0
+## 2.1.0
+
+### Minor Changes
+
+- 2d2ce8c: Enh: Added custom instance idle TTL support
+- 26d1070: Added support for instances hosted on alternate volumes
+- eb9d8a4: Feature: cloud volume support for instances
+
+### Patch Changes
+
+- Fix: Mothership notifications where not being sent out
+
+## 2.0.2
+
+### Patch Changes
+
+- Refactored mothership to TS using JSVM bundling
+
+## 2.0.1
+
+### Patch Changes
+
+- 05ae9b3: Fix: Discord health response code check
+
+## 2.0.0
 
 This release is a checkpoint release to jump over v1, which has been abandoned. v0 ended up becoming v2.